requirements.txt
numpy
pandas
scipy
pillow
scikit-learn
albumentations
opencv-python
jupyter 
notebook
tensorboard
matplotlib
timm
seaborn

torch
torchvision
pytorch-metric-learning
faiss-gpu

hydra-core
hydra_colorlog
